requirements.txt
docopt
pyyaml
django-ucamlookup
requests
